Airtable For Predictions/requirements.txt
Flask
numpy
pandas
sklearn
scikit-learn==1.2.2
streamlit
xgboost
fastapi
uvicorn
airtable-python-wrapper